move session folder to src
